backend/requirements.txt
alembic==1.14.1
amqp==5.3.1
async-timeout==5.0.1
backports.zoneinfo==0.2.1
bcrypt==5.0.0
bidict==0.23.1
billiard==4.2.4
blinker==1.8.2
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
dnspython==2.6.1
eventlet==0.39.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
h11==0.16.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
Mako==1.3.10
MarkupSafe==2.1.5
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.13.0
python-socketio==5.16.0
redis==6.1.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.13.2
tzdata==2025.3
vine==5.1.0
wcwidth==0.3.5
Werkzeug==3.0.6
wsproto==1.2.0
zipp==3.20.2
qrcode[pil]==7.4.2
python-barcode==0.14.0
Pillow==10.1.0